+++ lib/appsignal/instrumentation/decorators.ex
@@ -22,7 +22,7 @@ defmodule Appsignal.Instrumentation.Decorators do
   def instrument(namespace, body, %{module: module, name: name, arity: arity})
       when is_binary(namespace) do
     quote do
-      Appsignal.instrument(
+      Appsignal.Instrumentation.Helpers.instrument(
         "#{module_name(unquote(module))}.#{unquote(name)}/#{unquote(arity)}",
         fn span ->
           unquote(@span).set_namespace(span, unquote(namespace))
@@ -34,7 +34,7 @@ defmodule Appsignal.Instrumentation.Decorators do
 
   def instrument(body, %{module: module, name: name, arity: arity}) do
     quote do
-      Appsignal.instrument(
+      Appsignal.Instrumentation.Helpers.instrument(
         "#{module_name(unquote(module))}.#{unquote(name)}/#{unquote(arity)}",
         fn -> unquote(body) end
       )
@@ -43,7 +43,7 @@ defmodule Appsignal.Instrumentation.Decorators do
 
   def instrument(body, %{module: module, name: name}) do
     quote do
-      Appsignal.instrument(
+      Appsignal.Instrumentation.Helpers.instrument(
         "#{module_name(unquote(module))}.#{unquote(name)}",
         fn -> unquote(body) end
       )
